// Scout/src/types.ts
export interface ScoutEntry {
    id: string;
    scoutName: string;
    teamNumber: string;
    matchNumber: string;
    teamName: string;
    driverSkill: number;
    autonomousNotes: string;
    teleopNotes: string;
    generalNotes: string;
    timestamp: string;
    date: string;
}